Use openapi-generator to generate the go sdk
This is what a go api client looks like when developed by java
developers :smile:

Note that only the files not listed in `FILES` are maintained by us. An
easy way to look at them is to pull the project and run `make purge` to
blow away all the autogenerated stuff to see what you are able to
configure. Note that we can always override the templates if we want.

DBAAS-7829