Use deepmap/oapi-codegen to generate go client SDK
This is another client generator which consumes an openapi spec and
outputs the types and client methods necessary to interact with that
api. As compared to the openapi-generator, this version is stripped down
to only what you need.

This means there is much less functionality, with the tradeoff of
significantly less code to maintain. Looking at all the extra stuff from
the openapi-generator version, I'm not sure anything really of value was
lost. This SDK is extremely straight forward, and looks to do exactly
what we want of it - all while having a very flat implementation. When
compared to the layers of abstraction I saw over in the
openapi-generator, it makes me feel like I'm staring into the mines of
moria to get to the bottom of it.

I feel like from a maintainability standpoint, this client is the way to
go. In this case, less is more - we don't want to get fancy at this
layer, and just want a straightforward abstraction over the rest
endpoints.

DBAAS-7829