Use intuit/auto for automated release management
Some boilerplate here to do a basic release based on the generated code
that is sitting on main. I intend for us to come back through and add
another workflow which will handle automatically generating this client
and publishing to a `next` branch automatically - but for now this will
handle what happens whenenver something gets merged to main. That should
be enough for us to generate new clients by hand, open a PR, and get a
new release when it's accepted - just hopefully we won't have to do that
by hand for long.

DBAAS-7829