Remove unnecessary `needs` from release job in pipeline
Herp derp - that's what I get for copy / paste. Those jobs don't even
exist in this pipeline, so the workflow was invalid.

That's the hard part about working on a main-only pipeline, I can't
actually test without doing it for real (without using a dry-run and a
different `on`).

DBAAS-7829