Autogenerate SDK based on API mainline on next branch
I have a pretty low tolerance for manual tasks, so after generating the
SDK by hand a third time I figured enough was enough!

So this PR adds a new make target that retrieves the latest openapi spec
from the dev deployment (which represents the tip of the main branch for
the api project). That new target is incorporated into the `all` target
so that now it will generate using the latest spec whenever called.

Then, I added a new prereleaseBranch to the autorc (called `next`). This
way we can kick off the publish workflow on the next branch whenever we
like, and it will create a new pre-release build. Then, whenever we like
it and want to put it on the mainline for the SDK, we just make a PR
from `next` into `main`.

DBAAS-7829